Safari-611.1.10/LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-pt-PT-expected.txt

PASS navigator.language is "pt-pt"
PASS successfullyParsed is true

TEST COMPLETE
